lib: sbi_trap: Trap stack must be 16 bytes aligned for Zcherihybrid extension
Make sure the stack pointer is aligned to 16 bytes to as
`SBI_TRAP_CONTEXT_SIZE` is used to shift the stack pointer in trap
exception.

In integer pointer mode on zcherihybird assume the stack is aligned to
capability. scc instruction is used to initial data aligned to
capability, it would trigger alignement exception if the stack is not
aligned to capability.

Size of `struct sbi_trap_context` needs to be aligned to capability which
the size is 2x of `unsigned long`. Therefore, add an extra `unsigned long`
field is needed within `struct sbi_trap_context`. An 'unsidned long`
is added in `struct sbi_trap_info` for better code maintenance.

The non-cheri ABI specification also mentions the stack pointer must be
aligned to 16 bytes but it would not have an issue currently as
non-cheri ABI do not have any 16-byte (128-bit) aligned data type.