riscv_asm.h: Add Zcheripurecap ABI support
Defines macro to reprensent different usage for GPRs to improve the code
portability for Zcheripurecap ABI.
- REG(r) represents native GPRs with ABI name. e.g. REG(a0) is a0 on
  non-CHERI and ca0 on CHERI.
- PREG(r) represents native GPRs for storing pointer with ABI name. e.g.
  PREG(sp) is sp on non-CHERI and csp on CHERI.

Add REG, XREG aand PREG load/store macro to to improve the code
portability for CHERI extension. The native registers on the core
with CHERI extension is capability register.
- REG is the native register. XREG on non-CHERI and CREG on CHERI.
- XREG is the register for storing native word. XREG is same on both of
  non-CHERI and CHERI.
- PREG is the register for storing native pointer. XREG on non-CHERI and
  CREG on CHERI.

PC_PTR_L and PTR_L is added. For non-CHERI both are macro of lla
pseudo-instruction. For CHERI, PC_PTR_L is llc which use auipcc with
offset to derive the capability pointer from pcc. It can be used before
and after capability being initialized. The capability permission is
same as pcc. PTR_L can be used only after the capability initialization.
The capability pointer permission and bounds is set during capability
initialization.