lib: sbi_misaligned_ldst: Add Zcheripurecap ABI support for misaligned vector load/store
Create a capability pointer for handling misaligned vector load/store
which the permission is based on load or store emulations and the bounds
is based on the vector stride.

Remark: Current CHERI compiler has not yet supported the vector
load/store instructions with the correct capability operands. Please set
CC_SUPPORT_VECTOR=n explicitly as the auto detection in Makefile will
pass and set CC_SUPPORT_VECTOR=y as the vector instructions is supported
in standard RISC-V compilation.